// apps/admin/src/lib/api/generated/models/adminOrdersListParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen Platform Admin API
 * OpenAPI spec version: 0.1.0
 */
import type { AdminOrdersListPaymentStatus } from './adminOrdersListPaymentStatus';
import type { AdminOrdersListStatus } from './adminOrdersListStatus';

export type AdminOrdersListParams = {
customer_id?: string | null;
kitchen_id?: string | null;
community_id?: string | null;
status?: AdminOrdersListStatus;
payment_status?: AdminOrdersListPaymentStatus;
/**
 * @maxLength 200
 * @pattern ^[^\x00]*$
 */
q?: string;
sort?: string | null;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 * @maximum 10000
 */
offset?: number;
};